[TraceManager] Added TraceManager.Add(string message, TraceType state), TraceManager.AddAsync(string message, TraceType state), small performance optimization (#3134)

+++ External/Plugins/HaXeContext/Linters/DiagnosticsLinter.cs
@@ -148,7 +148,7 @@ namespace HaXeContext.Linters
                     }
                     break;
                 case HaxeCompleteStatus.ERROR:
-                    PluginBase.RunAsync(() => TraceManager.Add(hc.Errors, (int)TraceType.Error));
+                    PluginBase.RunAsync(() => TraceManager.Add(hc.Errors, TraceType.Error));
                     break;
             }
         }
